Run the proper kiwix script for the platform
The RACHEL-Plus script rachelkiwixstart.sh does not exist for the RACHEL-Pi and other installations. This will run the rachel-kiwix-start.pl perl script instead for RACHEL-Pi. The script execution is currently commented out as there is a bug with the killall kiwix-serve command inside of it hanging as well as updates to kiwix-serve itself for zim index management that are no longer necessary. That will be rewritten into an SH script.